# packages/compute-worker/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenReader Compute Worker API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/tts-playback/{sessionId}/audio": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "token",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sessionId",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Progressive MP3 audio stream",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "Progressive MP3 audio stream"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/v1/pdf-layout/operations": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "documentId": {
                    "type": "string",
                    "pattern": "^[a-f0-9]{64}$"
                  },
                  "namespace": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^[a-zA-Z0-9._-]{1,128}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "documentObjectKey": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2048
                  },
                  "replaceToken": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  }
                },
                "required": [
                  "documentId",
                  "namespace",
                  "documentObjectKey"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "opId": {
                      "type": "string"
                    },
                    "subject": {
                      "oneOf": [
                        {
                          "type": "object",
                          "properties": {
                            "kind": {
                              "type": "string",
                              "enum": [
                                "pdf_layout"
                              ]
                            },
                            "documentId": {
                              "type": "string"
                            },
                            "namespace": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "kind",
                            "documentId",
                            "namespace"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "object",
                          "properties": {
                            "kind": {
                              "type": "string",
                              "enum": [
                                "tts_playback"
                              ]
                            },
                            "documentId": {
                              "type": "string"
                            },
                            "sessionId": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "kind",
                            "documentId",
                            "sessionId"
                          ],
                          "additionalProperties": false
                        }
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "running",
                        "succeeded",
                        "failed"
                      ]
                    },
                    "queuedAt": {
                      "type": "number"
                    },
                    "updatedAt": {
                      "type": "number"
                    },
                    "startedAt": {
                      "type": "number"
                    },
                    "result": {},
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ],
                      "additionalProperties": false
                    },
                    "timing": {
                      "type": "object",
                      "properties": {
                        "queueWaitMs": {
                          "type": "number"
                        },
                        "s3FetchMs": {
                          "type": "number"
                        },
                        "computeMs": {
                          "type": "number"
                        }
                      },
                      "additionalProperties": false
                    },
                    "progress": {
                      "type": "object",
                      "properties": {
                        "totalPages": {
                          "type": "number"
                        },
                        "pagesParsed": {
                          "type": "number"
                        },
                        "currentPage": {
                          "type": "number"
                        },
                        "phase": {
                          "type": "string",
                          "enum": [
                            "infer",
                            "merge"
                          ]
                        }
                      },
                      "required": [
                        "totalPages",
                        "pagesParsed",
                        "phase"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "opId",
                    "subject",
                    "status",
                    "queuedAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/operations/{opId}/events": {
      "get": {
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            },
            "in": "query",
            "name": "sinceEventId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "opId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent ComputeOperationEvent stream",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "Server-sent ComputeOperationEvent stream"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}